Remove deprecated Java classes and adopt Kotlin versions
Deleted the old Java classes `FileHandler.java`, `Spell.java`, and `SpellCreator.java`. Renamed and refactored the newer Kotlin files to replace them.
